pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "price-bot"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0",
    "aiosqlite",
    "jinja2",
    "openai",
    "python-dotenv",
]

[project.scripts]
price-bot = "price_bot.bot.main:main"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
